LayoutTests/http/wpt/cache-storage/cache-quota-add.any-expected.txt

CONSOLE MESSAGE: Cache API operation failed: Quota exceeded
CONSOLE MESSAGE: Cache API operation failed: Quota exceeded

PASS Testing that cache.add checks against quota
PASS Testing that cache.addAll checks against quota
